[feat] 툭훈 UI 및 코인 연결
