+++ Assets/Plugins/Zenject/OptionalExtras/ReflectionBaking/Zenject-ReflectionBaking.csproj.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 9e00d9ec031f6bea295ef005ed419f3f
-PrefabImporter:
+guid: 91b0b4c77fdd1d24aa1af646e9d29960
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
