+++ Assets/Plugins/Zenject/OptionalExtras/Signals/Zenject-Signals.csproj.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: aebd7ef925dd4fe4690c3238e879f914
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
