+++ Assets/Programming/JYL/Test_Save/AutoSave.json
@@ -1,0 +1,24 @@
+{
+    "userId": "testUid123",
+    "playerName": "qwer",
+    "clanName": "qwer",
+    "saveSlotIndex": 0,
+    "time": {
+        "yearCycle": 1,
+        "season": 0,
+        "week": 1,
+        "playTick": 0,
+        "lastSaveUtcIso": "2025-09-04T03:39:47.1546825Z"
+    },
+    "currencies": {
+        "gold": 0,
+        "fame": 0,
+        "trainingCoin": 0
+    },
+    "buildings": [],
+    "athleteSaves": [],
+    "coachSaves": [],
+    "quests": [],
+    "achievements": [],
+    "encyclopedia": []
+}
