+++ Assets/Programming/JYL/Test_Save/AutoSave.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b32f34857aa34094796141b9ba55d009
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
